🚚 Move file related functions from helpers/security ti helpers/file
